fix: refactor Main and ResultNameDto
